# docs/openapi.json
{
  "components": {
    "schemas": {
      "ColumnModel": {
        "properties": {
          "fixed_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Encoded width in bytes, or null for variable-width types",
            "title": "Fixed Size"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nullable": {
            "title": "Nullable",
            "type": "boolean"
          },
          "primary_key": {
            "title": "Primary Key",
            "type": "boolean"
          },
          "type": {
            "enum": [
              "INTEGER",
              "FLOAT",
              "BOOLEAN",
              "TEXT"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type",
          "nullable",
          "primary_key",
          "fixed_size"
        ],
        "title": "ColumnModel",
        "type": "object"
      },
...
      "DatabaseDetail": {
        "properties": {
          "database_id": {
            "title": "Database Id",
            "type": "string"
          },
          "file_size_bytes": {
            "title": "File Size Bytes",
            "type": "integer"
          },
          "format_version": {
            "title": "Format Version",
            "type": "integer"
          },
          "free_list_head": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free List Head"
          },
          "heap_page_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Heap Page Ids",
            "type": "array"
          },
          "page_count": {
            "title": "Page Count",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Live rows; null when no table is defined yet",
            "title": "Row Count"
          },
          "schema": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SchemaModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "schema_page_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Schema Page Ids",
            "type": "array"
          },
          "stats": {
            "$ref": "#/components/schemas/PagerStatsModel"
          },
          "table_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table Name"
          },
          "trace_level": {
            "title": "Trace Level",
            "type": "string"
          }
        },
        "required": [
          "database_id",
          "page_size",
          "page_count",
          "file_size_bytes",
          "format_version",
          "table_name",
          "row_count",
          "heap_page_ids",
          "schema_page_ids",
          "free_list_head",
          "stats",
          "trace_level"
        ],
        "title": "DatabaseDetail",
        "type": "object"
      },
...
      "EventCategory": {
        "description": "Which subsystem produced an event.\n\nConsumers filter on this, so the full set is declared up front even though\nMilestone 1 only emits the first three.  A string enum keeps the value\nself-describing when it reaches JSON.",
        "enum": [
          "lifecycle",
          "storage",
          "record",
          "parser",
          "operator",
          "catalog",
          "index",
          "planner",
          "buffer_pool",
          "transaction",
          "wal",
          "recovery",
          "lock",
          "mvcc"
        ],
        "title": "EventCategory",
        "type": "string"
      },
      "EventsResponse": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/TraceRecordModel"
            },
            "title": "Events",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          },
          "stats": {
            "$ref": "#/components/schemas/TraceStatsModel"
          }
        },
        "required": [
          "events",
          "stats",
          "page"
        ],
        "title": "EventsResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageInfo": {
        "description": "Cursor pagination for the event stream.",
        "properties": {
          "after_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Seq"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as after_seq to continue; null when caught up",
            "title": "Next Cursor"
          },
          "returned": {
            "title": "Returned",
            "type": "integer"
          }
        },
        "required": [
          "after_seq",
          "returned",
          "next_cursor",
          "has_more"
        ],
        "title": "PageInfo",
        "type": "object"
      },
...
      "PagerStatsModel": {
        "description": "Cumulative I/O counters since the database handle was opened.",
        "properties": {
          "allocations": {
            "title": "Allocations",
            "type": "integer"
          },
          "bytes_read": {
            "title": "Bytes Read",
            "type": "integer"
          },
          "bytes_written": {
            "title": "Bytes Written",
            "type": "integer"
          },
          "frees": {
            "title": "Frees",
            "type": "integer"
          },
          "page_reads": {
            "title": "Page Reads",
            "type": "integer"
          },
          "page_writes": {
            "title": "Page Writes",
            "type": "integer"
          },
          "read_time_ns": {
            "title": "Read Time Ns",
            "type": "integer"
          },
          "recycled_allocations": {
            "title": "Recycled Allocations",
            "type": "integer"
          },
          "syncs": {
            "title": "Syncs",
            "type": "integer"
          },
          "write_time_ns": {
            "title": "Write Time Ns",
            "type": "integer"
          }
        },
        "required": [
          "page_reads",
          "page_writes",
          "allocations",
          "recycled_allocations",
          "frees",
          "syncs",
          "bytes_read",
          "bytes_written",
          "read_time_ns",
          "write_time_ns"
        ],
        "title": "PagerStatsModel",
        "type": "object"
      },
...
      "SchemaModel": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnModel"
            },
            "title": "Columns",
            "type": "array"
          },
          "fixed_row_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Constant encoded row size, or null if any column varies",
            "title": "Fixed Row Size"
          },
          "null_bitmap_size": {
            "description": "Bytes of null bitmap per record",
            "title": "Null Bitmap Size",
            "type": "integer"
          }
        },
        "required": [
          "columns",
          "null_bitmap_size",
          "fixed_row_size"
        ],
        "title": "SchemaModel",
        "type": "object"
      },
      "SetTraceLevelRequest": {
        "additionalProperties": false,
        "properties": {
          "level": {
            "enum": [
              "OFF",
              "SUMMARY",
              "OPERATOR",
              "STORAGE",
              "VERBOSE"
            ],
            "title": "Level",
            "type": "string"
          }
        },
        "required": [
          "level"
        ],
        "title": "SetTraceLevelRequest",
        "type": "object"
      },
...
      "TraceLevelResponse": {
        "properties": {
          "level": {
            "enum": [
              "OFF",
              "SUMMARY",
              "OPERATOR",
              "STORAGE",
              "VERBOSE"
            ],
            "title": "Level",
            "type": "string"
          },
          "stats": {
            "$ref": "#/components/schemas/TraceStatsModel"
          }
        },
        "required": [
          "level",
          "stats"
        ],
        "title": "TraceLevelResponse",
        "type": "object"
      },
      "TraceRecordModel": {
        "description": "One diagnostic event with the envelope the tracer stamped on it.",
        "properties": {
          "category": {
            "enum": [
              "lifecycle",
              "storage",
              "record",
              "parser",
              "operator",
              "catalog",
              "index",
              "planner",
              "buffer_pool",
              "transaction",
              "wal",
              "recovery",
              "lock",
              "mvcc"
            ],
            "title": "Category",
            "type": "string"
          },
          "event": {
            "additionalProperties": true,
            "description": "Flat payload; fields depend on event_type",
            "title": "Event",
            "type": "object"
          },
          "event_type": {
            "description": "Event class name, e.g. 'PageReadEvent'",
            "title": "Event Type",
            "type": "string"
          },
          "level": {
            "enum": [
              "OFF",
              "SUMMARY",
              "OPERATOR",
              "STORAGE",
              "VERBOSE"
            ],
            "title": "Level",
            "type": "string"
          },
          "seq": {
            "description": "Monotonic per database; also the pagination cursor",
            "title": "Seq",
            "type": "integer"
          },
          "timestamp_ns": {
            "title": "Timestamp Ns",
            "type": "integer"
          }
        },
        "required": [
          "seq",
          "timestamp_ns",
          "category",
          "level",
          "event_type",
          "event"
        ],
        "title": "TraceRecordModel",
        "type": "object"
      },
      "TraceStatsModel": {
        "description": "Retention state, so a client can tell when it has missed events.",
        "properties": {
          "capacity": {
            "title": "Capacity",
            "type": "integer"
          },
          "dropped": {
            "description": "Events evicted from the ring buffer before being read",
            "title": "Dropped",
            "type": "integer"
          },
          "level": {
            "enum": [
              "OFF",
              "SUMMARY",
              "OPERATOR",
              "STORAGE",
              "VERBOSE"
            ],
            "title": "Level",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total_recorded": {
            "title": "Total Recorded",
            "type": "integer"
          }
        },
        "required": [
          "capacity",
          "size",
          "total_recorded",
          "dropped",
          "level"
        ],
        "title": "TraceStatsModel",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "HTTP and WebSocket access to a running ChenDB engine, for the Visual Database\nExplorer.\n\nEverything served here is real engine state read from a real database file.\nEndpoints for features that do not exist yet are absent rather than stubbed:\ncheck `features` on `/health` to see what this build supports.",
    "title": "ChenDB Engine API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/databases/{database_id}": {
      "delete": {
        "operationId": "delete_database_api_v1_databases__database_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "database_id",
            "required": true,
            "schema": {
              "title": "Database Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a database file",
        "tags": [
          "databases"
        ]
      },
      "get": {
        "operationId": "get_database_detail_api_v1_databases__database_id__get",
        "parameters": [
          {
            "description": "Workspace-relative database id",
            "in": "path",
            "name": "database_id",
            "required": true,
            "schema": {
              "description": "Workspace-relative database id",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Database Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Database detail",
        "tags": [
          "databases"
        ]
      }
    },
    "/api/v1/databases/{database_id}/events": {
      "delete": {
        "operationId": "clear_events_api_v1_databases__database_id__events_delete",
        "parameters": [
          {
            "description": "Workspace-relative database id",
            "in": "path",
            "name": "database_id",
            "required": true,
            "schema": {
              "description": "Workspace-relative database id",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Database Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceLevelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear retained events",
        "tags": [
          "diagnostics"
        ]
      },
      "get": {
        "description": "Page through retained events by sequence number.\n\n``after_seq`` is a stable cursor: sequence numbers are assigned\nmonotonically by the tracer, so paging cannot skip or repeat an event even\nwhile new ones arrive.",
        "operationId": "list_events_api_v1_databases__database_id__events_get",
        "parameters": [
          {
            "description": "Workspace-relative database id",
            "in": "path",
            "name": "database_id",
            "required": true,
            "schema": {
              "description": "Workspace-relative database id",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Database Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_seq",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Seq"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 2000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/EventCategory"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recent events",
        "tags": [
          "diagnostics"
        ]
      }
    },
...
    "/api/v1/databases/{database_id}/trace": {
      "get": {
        "operationId": "get_trace_level_api_v1_databases__database_id__trace_get",
        "parameters": [
          {
            "description": "Workspace-relative database id",
            "in": "path",
            "name": "database_id",
            "required": true,
            "schema": {
              "description": "Workspace-relative database id",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Database Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceLevelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get the trace level",
        "tags": [
          "diagnostics"
        ]
      },
      "put": {
        "operationId": "set_trace_level_api_v1_databases__database_id__trace_put",
        "parameters": [
          {
            "description": "Workspace-relative database id",
            "in": "path",
            "name": "database_id",
            "required": true,
            "schema": {
              "description": "Workspace-relative database id",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Database Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTraceLevelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceLevelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set the trace level",
        "tags": [
          "diagnostics"
        ]
      }
    },
...
  }
}